# app/tests/backend/agent_api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Dispatch agent API",
    "version": "1",
    "description": "Read a DSP's drivers, routes, timecards, meal breaks, DVIC and weekly scorecards, as far as the key may read them there. Every answer names the DSP, the days and the driver it understood, and which days each source has. An answer read from a feature the DSP has switched off, by a key that bypasses features, names it under `bypassed`. Times are the DSP's own."
  },
  "servers": [
    {
      "url": "https://dispatch.example.com"
    }
  ],
  "security": [
    {
      "key": []
    }
  ],
  "components": {
    "securitySchemes": {
      "key": {
        "type": "http",
        "scheme": "bearer",
        "description": "An agent key made on the Agents page, or an OAuth access token issued by Sign in with Dispatch."
      }
    }
  },
  "paths": {
...
    "/api/v1/routes": {
      "get": {
        "operationId": "routes",
        "summary": "Routes over a period",
        "description": "Routes for one driver or everyone over a period, with package and stop totals, departure and end.",
        "parameters": [
          {
            "name": "dsp",
            "in": "query",
            "required": false,
            "description": "The DSP, by name. Leave it out when the key reaches one DSP.",
            "schema": {
              "type": "string",
              "maxLength": 200
            }
          },
          {
            "name": "period",
            "in": "query",
            "required": false,
            "description": "Days in the DSP's time: last week, last 14 days or 2026-W39. Defaults to yesterday.",
            "schema": {
              "type": "string",
              "maxLength": 200
            }
          },
          {
            "name": "date",
            "in": "query",
            "required": false,
            "description": "One day: today, yesterday, last night or 2026-09-28.",
            "schema": {
              "type": "string",
              "maxLength": 200
            }
          },
          {
            "name": "from",
            "in": "query",
            "required": false,
            "description": "The first day, as 2026-09-01, with `to`; instead of `period`.",
            "schema": {
              "type": "string",
              "maxLength": 200
            }
          },
          {
            "name": "to",
            "in": "query",
            "required": false,
            "description": "The last day, as 2026-09-30, with `from`.",
            "schema": {
              "type": "string",
              "maxLength": 200
            }
          },
          {
            "name": "driver",
            "in": "query",
            "required": false,
            "description": "A driver as the user named them: a name or part of one, a Driver Match code, a Paycom employee code or an Amazon transporter ID.",
            "schema": {
              "type": "string",
              "maxLength": 200
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "description": "The most rows to return, 1 to 500.",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 500
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "description": "The next_cursor an earlier answer gave, for its next page.",
            "schema": {
              "type": "string",
              "maxLength": 200
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The answer.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "final": {
                      "type": "boolean"
                    },
                    "coverage": {
                      "type": "object",
                      "properties": {
                        "status": {
                          "type": "string",
                          "enum": [
                            "complete",
                            "partial",
                            "missing",
                            "unavailable"
                          ]
                        },
                        "enabled": {
                          "type": "boolean"
                        },
                        "collected": {
                          "type": "integer",
                          "minimum": 0
                        },
                        "of": {
                          "type": "integer",
                          "minimum": 0
                        },
                        "missing": {
                          "type": "array",
                          "items": {
                            "type": "string"
                          }
                        },
                        "snapshots": {
                          "type": "array",
                          "items": {
                            "type": "string"
                          }
                        },
                        "not_posted_yet": {
                          "type": "array",
                          "items": {
                            "type": "string"
                          }
                        }
                      },
                      "required": [
                        "status"
                      ]
                    },
                    "totals": {
                      "anyOf": [
                        {
                          "type": "object",
                          "additionalProperties": {
                            "anyOf": [
                              {
                                "type": "number"
                              },
                              {
                                "type": "null"
                              }
                            ]
                          }
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "routes": {
                      "type": "object",
                      "properties": {
                        "columns": {
                          "type": "array",
                          "items": {
                            "type": "string",
                            "enum": [
                              "date",
                              "route",
                              "driver",
                              "packages",
                              "delivered",
                              "undeliverable",
                              "stops",
                              "completed",
                              "departed",
                              "ended"
                            ]
                          }
                        },
                        "rows": {
                          "type": "array",
                          "items": {
                            "type": "array",
                            "items": {}
                          }
                        },
                        "page": {
                          "type": "object",
                          "properties": {
                            "returned": {
                              "type": "integer",
                              "minimum": 0
                            },
                            "total": {
                              "type": "integer",
                              "minimum": 0
                            },
                            "next_cursor": {
                              "type": "string"
                            }
                          },
                          "required": [
                            "returned",
                            "total"
                          ]
                        },
                        "note": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "columns",
                        "rows"
                      ]
                    },
                    "understood": {
                      "type": "object",
                      "properties": {
                        "dsp": {
                          "type": "string"
                        },
                        "today": {
                          "type": "string"
                        },
                        "period": {
                          "type": "string"
                        },
                        "from": {
                          "type": "string"
                        },
                        "to": {
                          "type": "string"
                        },
                        "days": {
                          "type": [
                            "integer",
                            "string"
                          ]
                        },
                        "driver": {
                          "anyOf": [
                            {
                              "type": "string"
                            },
                            {
                              "type": "object",
                              "properties": {
                                "code": {
                                  "type": "string"
                                },
                                "name": {
                                  "type": "string"
                                }
                              },
                              "required": [
                                "code",
                                "name"
                              ]
                            }
                          ]
                        },
                        "week": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "dsp",
                        "today"
                      ]
                    },
                    "note": {
                      "type": "string"
                    },
                    "bypassed": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    },
                    "not_allowed": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    },
                    "switched_off": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    }
                  },
                  "required": [
                    "final",
                    "coverage",
                    "totals",
                    "understood"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Something unclear; message says what and choices what it could mean.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "choices": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "No key, or a key that is revoked, expired or not for this Dispatch.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "choices": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          },
          "403": {
            "description": "Not for this key here: not_allowed when the key may not read it at the DSP, source_off when the DSP has the feature switched off.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "choices": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Nothing by that name, or data not collected or posted yet.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "choices": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          },
          "422": {
            "description": "The query needs too much database work; narrow the days or filters.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "choices": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          },
          "429": {
            "description": "Too many calls this minute; wait for Retry-After seconds.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "choices": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            },
            "headers": {
              "Retry-After": {
                "description": "Seconds to wait before retrying.",
                "schema": {
                  "type": "integer",
                  "minimum": 1
                }
              }
            }
          },
          "500": {
            "description": "Dispatch could not answer.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "choices": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          },
          "503": {
            "description": "Dispatch is temporarily busy or unavailable.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "choices": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            },
            "headers": {
              "Retry-After": {
                "description": "Seconds to wait before retrying.",
                "schema": {
                  "type": "integer",
                  "minimum": 1
                }
              }
            }
          }
        }
      }
    },
...
  }
}